Make sure that there's a <td> inside the highlight row
The <tr> isn't displayed if there isn't a <td> inside it.  This puts it
there.  It also makes the assumption that there's one <tr> that has a
<td colspan="X"> inside it and uses that to figure out how many columns
there are.  Not the best way, but it works since we know that Django
currently adds that at the top of the table.